# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nexo API",
        "description": "Fiscal centralizing API for integration between sales systems, government and accounting platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "https://nexo.test",
            "description": "Herd Local Server (HTTPS)"
        },
        {
            "url": "https://backendnexo.shop",
            "description": "Mounted Server"
        }
    ],
    "paths": {
...
        "/api/sales": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Submit a sale for processing",
                "description": "Accepts a sale payload and dispatches it for asynchronous processing. Step 1: creates a Xero Invoice (ACCREC). Step 2 (only if Xero succeeds): sends the invoice to TaxCore for fiscal signing. Returns immediately with a sale ID and pending status. Poll GET /api/sales/{id} for the result.",
                "operationId": "storeSale",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoiceType",
                                    "transactionType",
                                    "buyer",
                                    "items",
                                    "payment"
                                ],
                                "properties": {
                                    "invoiceType": {
                                        "description": "0=Normal, 1=Proforma, 2=Copy, 3=Training, 4=Advance",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    },
                                    "transactionType": {
                                        "description": "0=Sale, 1=Refund",
                                        "type": "integer",
                                        "example": 0,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "cashier": {
                                        "type": "string",
                                        "example": "Juan Pérez",
                                        "nullable": true,
                                        "maxLength": 50
                                    },
                                    "dueDate": {
                                        "description": "Invoice due date sent to Xero (YYYY-MM-DD). Defaults to today if omitted.",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-29",
                                        "nullable": true
                                    },
                                    "referentDocumentNumber": {
                                        "description": "Required for Refund (transactionType=1) or Copy/Advance (invoiceType=2 or 4).",
                                        "type": "string",
                                        "example": null,
                                        "nullable": true,
                                        "maxLength": 50
                                    },
                                    "buyer": {
                                        "required": [
                                            "name"
                                        ],
                                        "properties": {
                                            "name": {
                                                "description": "Buyer name used as Contact in Xero Invoice.",
                                                "type": "string",
                                                "example": "Empresa ABC d.o.o.",
                                                "maxLength": 255
                                            },
                                            "id": {
                                                "description": "Buyer tax ID sent to TaxCore as buyerId.",
                                                "type": "string",
                                                "example": "101234567",
                                                "nullable": true,
                                                "maxLength": 20
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "code",
                                                "name",
                                                "quantity",
                                                "unitPrice",
                                                "totalAmount",
                                                "labels",
                                                "accountCode"
                                            ],
                                            "properties": {
                                                "code": {
                                                    "description": "Xero ItemCode — must match an existing Xero item.",
                                                    "type": "string",
                                                    "example": "PROD-001",
                                                    "maxLength": 30
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Producto A",
                                                    "maxLength": 2048
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2
                                                },
                                                "unitPrice": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "totalAmount": {
                                                    "description": "Sent to TaxCore.",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "labels": {
                                                    "description": "TaxCore fiscal tax labels.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "minItems": 1
                                                },
                                                "accountCode": {
                                                    "description": "Xero account code for the line item.",
                                                    "type": "string",
                                                    "example": "200",
                                                    "maxLength": 10
                                                },
                                                "gtin": {
                                                    "description": "Optional GTIN for TaxCore.",
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    },
                                    "payment": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "amount",
                                                "paymentType"
                                            ],
                                            "properties": {
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "paymentType": {
                                                    "description": "0=Other, 1=Cash, 2=Card, 3=Check, 4=WireTransfer, 5=Voucher, 6=MobileMoney",
                                                    "type": "integer",
                                                    "example": 1,
                                                    "enum": [
                                                        0,
                                                        1,
                                                        2,
                                                        3,
                                                        4,
                                                        5,
                                                        6
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Sale accepted for processing.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale submitted for processing."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid connector token."
                    },
                    "422": {
                        "description": "Validation error — missing or invalid fields."
                    }
                },
                "security": [
                    {
                        "connectorToken": []
                    }
                ]
            }
        },
...
        "/api/integrations/taxcore/invoices/{id}": {
            "get": {
                "tags": [
                    "TaxCore Integration"
                ],
                "summary": "Get a single fiscalized invoice (read from local DB)",
                "description": "Returns the full Sale record by its internal ID, including the original request payload, the TaxCore signed response and the Xero response. Reads from the `sales` table. organization_id is required and the sale is verified to belong to that organization. No authentication required.",
                "operationId": "getTaxcoreInvoice",
                "parameters": [
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "ID of the organization that owns the invoice.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Internal Sale ID.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice retrieved successfully."
                    },
                    "404": {
                        "description": "Invoice not found for this organization."
                    },
                    "422": {
                        "description": "Validation error — organization_id is required."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "agentToken": {
                "type": "http",
                "description": "Bearer token issued by POST /api/agent/token. Used by the nexo-agent daemon to authenticate all agent endpoints.",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT issued by POST /api/auth/login. Pass as: Authorization: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "connectorToken": {
                "type": "http",
                "description": "Bearer token issued when a Connector is created. Only visible at creation time.",
                "scheme": "bearer"
            }
        }
    },
...
}